Movimiento de DTOs a paquete específico
Movimiento

+++ Training_FE/src/main/java/io/interfaz/training/pojos/Orders.java
@@ -6,6 +6,7 @@ package io.interfaz.training.pojos;
 
 
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import lombok.AllArgsConstructor;
@@ -24,7 +25,7 @@ public class Orders {
 	private int id;
 	private Date purchaseDate;
 	private int customerId;
-	private double subtotal;
-	private double iva;
-	private double total;
+	private BigDecimal subtotal;
+	private BigDecimal iva;
+	private BigDecimal total;
 }
